Revert "Email confirmation (#1330)"
This reverts commit 1fa23435862272027170083c985eaf35d1d649ad.
