+++ client/src/__tests__/ProfilePage.test.tsx
@@ -8,8 +8,6 @@ import { ProfilePage } from '../pages/profile';
 jest.mock('next/config', () => () => ({}));
 jest.mock('api', () => ({
   ProfileApi: jest.fn(),
-  UsersNotificationsApi: jest.fn(),
-  NotificationsApi: jest.fn(),
   CoursesApi: jest.fn(),
 }));
