+++ client/src/components/Profile/ContactsCard.tsx
@@ -10,8 +10,6 @@ import { CheckboxChangeEvent } from 'antd/lib/checkbox';
 const { Text } = Typography;
 
 import { ContactsOutlined } from '@ant-design/icons';
-import { NotificationChannel } from 'modules/Notifications/services/notifications';
-import { EmailConfirmation } from './EmailConfirmation';
 
 type Props = {
   data: Contacts;
@@ -19,18 +17,6 @@ type Props = {
   permissionsSettings?: ConfigurableProfilePermissions;
   onPermissionsSettingsChange: (event: CheckboxChangeEvent, settings: ChangedPermissionsSettings) => void;
   onProfileSettingsChange: (event: any, path: string) => void;
-  connections: Partial<
-    Record<
-      NotificationChannel,
-      | {
-          value: string;
-          enabled: boolean;
-          lastLinkSentAt?: string;
-        }
-      | undefined
-    >
-  >;
-  sendConfirmationEmail: () => void;
 };
 
 type Contact = { name: string; value: string | null; key: string };
@@ -74,20 +60,13 @@ class ContactsCard extends React.Component<Props> {
       !isEqual(nextProps.permissionsSettings?.isSkypeVisible, isSkypeVisible) ||
       !isEqual(nextProps.permissionsSettings?.isContactsNotesVisible, isContactsNotesVisible) ||
       !isEqual(nextProps.permissionsSettings?.isLinkedInVisible, isLinkedInVisible) ||
-      !isEqual(nextProps.isEditingModeEnabled, this.props.isEditingModeEnabled) ||
-      !isEqual(nextProps.connections, this.props.connections)
+      !isEqual(nextProps.isEditingModeEnabled, this.props.isEditingModeEnabled)
     );
   };
 
   render() {
-    const {
-      isEditingModeEnabled,
-      permissionsSettings,
-      onPermissionsSettingsChange,
-      onProfileSettingsChange,
-      connections,
-      sendConfirmationEmail,
-    } = this.props;
+    const { isEditingModeEnabled, permissionsSettings, onPermissionsSettingsChange, onProfileSettingsChange } =
+      this.props;
     const { email, epamEmail, telegram, phone, skype, notes, linkedIn } = this.props.data;
     const contacts = [
       { name: 'EPAM E-mail', value: epamEmail, key: 'epamEmail' },
@@ -138,15 +117,7 @@ class ContactsCard extends React.Component<Props> {
                 <List.Item>
                   <Text strong>{name}:</Text>{' '}
                   {key !== 'linkedIn' ? (
-                    <>
-                      {value}
-                      {key === 'email' && (!connections.email || !connections.email.enabled) ? (
-                        <EmailConfirmation
-                          connection={connections.email}
-                          sendConfirmationEmail={sendConfirmationEmail}
-                        />
-                      ) : null}
-                    </>
+                    value
                   ) : (
                     <a target="__blank" href={value!}>
                       {value}
