+++ client/src/components/Profile/__test__/ContactsCard.test.tsx
@@ -21,8 +21,6 @@ describe('ContactsCard', () => {
           isEditingModeEnabled={false}
           onPermissionsSettingsChange={jest.fn()}
           onProfileSettingsChange={jest.fn()}
-          sendConfirmationEmail={jest.fn()}
-          connections={{}}
         />,
       );
       expect(shallowToJson(wrapper)).toMatchSnapshot();
@@ -42,8 +40,6 @@ describe('ContactsCard', () => {
           isEditingModeEnabled={true}
           onPermissionsSettingsChange={jest.fn()}
           onProfileSettingsChange={jest.fn()}
-          sendConfirmationEmail={jest.fn()}
-          connections={{}}
         />,
       );
       expect(shallowToJson(wrapper)).toMatchSnapshot();
