+++ client/src/pages/profile/index.tsx
@@ -39,7 +39,6 @@ import { withGoogleMaps } from 'components/withGoogleMaps';
 
 import { CheckboxChangeEvent } from 'antd/lib/checkbox';
 import { checkIsProfileOwner } from 'utils/profile-check';
-import { NotificationChannel, NotificationsService } from 'modules/Notifications/services/notifications';
 
 type Props = {
   router: NextRouter;
@@ -56,16 +55,6 @@ type State = {
   isEditingModeEnabled: boolean;
   isInitialPermissionsSettingsChanged: boolean;
   isInitialProfileSettingsChanged: boolean;
-  connections: Partial<
-    Record<
-      NotificationChannel,
-      | {
-          value: string;
-          enabled: boolean;
-        }
-      | undefined
-    >
-  >;
 };
 
 export type ChangedPermissionsSettings = {
@@ -83,7 +72,6 @@ export class ProfilePage extends React.Component<Props, State> {
     isEditingModeEnabled: false,
     isInitialPermissionsSettingsChanged: false,
     isInitialProfileSettingsChanged: false,
-    connections: {},
   };
 
   private onPermissionsSettingsChange = async (
@@ -168,7 +156,6 @@ export class ProfilePage extends React.Component<Props, State> {
   };
 
   private userService = new UserService();
-  private notificationsService = new NotificationsService();
   private coursesService = new CoursesService();
 
   private getCoursesInfo = async (profile: ProfileInfo) =>
@@ -204,10 +191,7 @@ export class ProfilePage extends React.Component<Props, State> {
 
     try {
       const githubId = router.query ? (router.query.githubId as string) : undefined;
-      const [profile, connections] = await Promise.all([
-        this.userService.getProfileInfo(githubId),
-        this.notificationsService.getUserConnections().catch(() => []),
-      ]);
+      const profile = await this.userService.getProfileInfo(githubId);
 
       const coursesInfo = await this.getCoursesInfo(profile);
 
@@ -236,7 +220,6 @@ export class ProfilePage extends React.Component<Props, State> {
         initialPermissionsSettings,
         isEditingModeEnabled,
         initialProfileSettings,
-        connections: connections as State['connections'],
       });
     } catch (e) {
       this.setState({
@@ -260,14 +243,6 @@ export class ProfilePage extends React.Component<Props, State> {
     message.error('Error has occured. Please check your connection and try again');
   }
 
-  private sendEmailConfirmationLink = async () => {
-    try {
-      await this.userService.sendEmailConfirmationLink();
-    } catch (e) {
-      message.error('Error has occured. Please try again later');
-    }
-  };
-
   private saveProfile = async () => {
     const { profile, isInitialPermissionsSettingsChanged, isInitialProfileSettingsChanged } = this.state;
 
@@ -284,7 +259,6 @@ export class ProfilePage extends React.Component<Props, State> {
           isPermissionsSettingsChanged: isInitialPermissionsSettingsChanged,
           isProfileSettingsChanged: isInitialProfileSettingsChanged,
         });
-        const connections = await this.notificationsService.getUserConnections().catch(() => []);
 
         const initialPermissionsSettings = permissionsSettings ? cloneDeep(permissionsSettings) : null;
         const initialProfileSettings = profile ? cloneDeep(profile) : null;
@@ -294,7 +268,6 @@ export class ProfilePage extends React.Component<Props, State> {
           initialProfileSettings,
           isInitialPermissionsSettingsChanged: false,
           isInitialProfileSettingsChanged: false,
-          connections: connections as State['connections'],
         });
         this.onSaveSuccess();
       } catch (e) {
@@ -346,7 +319,6 @@ export class ProfilePage extends React.Component<Props, State> {
       isInitialPermissionsSettingsChanged,
       isInitialProfileSettingsChanged,
       isProfileOwner,
-      connections,
     } = this.state;
     const isEditingModeVisible = initialPermissionsSettings && isEditingModeEnabled ? isEditingModeEnabled : false;
     const isSaveButtonVisible = isInitialPermissionsSettingsChanged || isInitialProfileSettingsChanged;
@@ -396,8 +368,6 @@ export class ProfilePage extends React.Component<Props, State> {
           permissionsSettings={profile.permissionsSettings}
           onPermissionsSettingsChange={this.onPermissionsSettingsChange}
           onProfileSettingsChange={this.onProfileSettingsChange}
-          connections={connections}
-          sendConfirmationEmail={this.sendEmailConfirmationLink}
         />
       ),
       profile?.discord !== undefined && <DiscordCard data={profile.discord} isProfileOwner={isProfileOwner} />,
