+++ nestjs/src/auth/auth.controller.ts
@@ -8,7 +8,6 @@ import { AuthService, CurrentRequest } from './auth.service';
 import { JWT_COOKIE_NAME } from './constants';
 import { AuthConnectionDto } from './dto/AuthConnectionDto';
 import { GithubStrategy } from './strategies/github.strategy';
-import * as dayjs from 'dayjs';
 
 const isDev = process.env.NODE_ENV !== 'production';
 const twoDaysMs = 1000 * 60 * 60 * 24 * 2;
@@ -58,10 +57,7 @@ export class AuthController {
   @UseGuards(DefaultGuard, RoleGuard)
   @RequiredRoles([Role.Admin])
   async createConnectLinkViaGithub(@Body() dto: AuthConnectionDto) {
-    const link = await this.githubStrategy.getAuthorizeUrl({
-      data: dto,
-      expires: dayjs().add(1, 'hour').toISOString(),
-    });
+    const link = await this.githubStrategy.getAuthorizeUrl(dto);
     return {
       link,
     };
