+++ nestjs/src/auth/auth.service.ts
@@ -6,7 +6,8 @@ import { InjectRepository } from '@nestjs/typeorm';
 import type { Request } from 'express';
 import { customAlphabet } from 'nanoid/async';
 import type { Profile } from 'passport';
-import { MoreThanOrEqual, Repository } from 'typeorm';
+import { UserNotificationsService } from '../users/users.notifications.service';
+import { MoreThanOrEqual } from 'typeorm';
 import { ConfigService } from '../config';
 import { CourseTasksService } from '../courses';
 import { UsersService } from '../users/users.service';
@@ -14,20 +15,12 @@ import { AuthUser } from './auth-user.model';
 import { AuthRepository } from './auth.repository';
 import { JwtService } from './jwt.service';
 import { lastValueFrom } from 'rxjs';
-import * as dayjs from 'dayjs';
-import { NotificationUserConnection } from '@entities/notificationUserConnection';
 
 const nanoid = customAlphabet('1234567890abcdef', 10);
 
 export type CurrentRequest = Request & {
   user: AuthUser;
   loginState?: LoginData;
-};
-
-export type LoginStateParams = {
-  userId?: number;
-  expires?: string;
-  data: LoginData;
 };
 
 @Injectable()
@@ -41,8 +34,7 @@ export class AuthService {
     readonly configService: ConfigService,
     @InjectRepository(AuthRepository)
     private readonly authRepository: AuthRepository,
-    @InjectRepository(NotificationUserConnection)
-    private notificationUserConnectionRepository: Repository<NotificationUserConnection>,
+    readonly userNotificationsService: UserNotificationsService,
     private httpService: HttpService,
   ) {
     this.admins = configService.users.admins;
@@ -101,40 +93,24 @@ export class AuthService {
     return this.jwtService.createToken(req.user);
   }
 
-  public async createLoginState(params: LoginStateParams) {
+  public async createLoginState(data: LoginData) {
     const id = await nanoid();
-    const { data, expires, userId } = params;
 
     await this.authRepository.save({
       id,
       data,
-      userId,
-      expires,
     });
 
     return id;
   }
 
-  public getLoginStateById(id: string) {
+  public getLoginState(id: string) {
+    const date = new Date();
+    date.setHours(date.getHours() - 1);
     return this.authRepository.findOne({
       where: {
         id,
-        expires: MoreThanOrEqual(dayjs().toISOString()),
-      },
-      order: {
-        createdDate: 'DESC',
-      },
-    });
-  }
-
-  public getLoginStateByUserId(id: number) {
-    return this.authRepository.findOne({
-      where: {
-        userId: id,
-        expires: MoreThanOrEqual(dayjs().toISOString()),
-      },
-      order: {
-        createdDate: 'DESC',
+        createdDate: MoreThanOrEqual(date.toISOString()),
       },
     });
   }
@@ -150,13 +126,12 @@ export class AuthService {
   public async onConnectionComplete(loginData: LoginData, userId: number) {
     const { channelId, externalId } = loginData;
 
-    this.notificationUserConnectionRepository.save({
+    await this.userNotificationsService.saveUserConnection({
       channelId,
       enabled: true,
       externalId,
       userId,
     });
-
     const { restApiKey, restApiUrl } = this.configService.awsServices;
 
     if (channelId === 'telegram') {
