+++ nestjs/src/certificates/certificates.controller.ts
@@ -2,17 +2,14 @@ import { Body, Controller, Get, NotFoundException, Param, Post, Res, UseGuards }
 import { ApiOperation, ApiTags } from '@nestjs/swagger';
 import { Response } from 'express';
 import { DefaultGuard, RequiredRoles, Role, RoleGuard } from 'src/auth';
-import { UserNotificationsService } from 'src/usersNotifications/users.notifications.service';
+import { NotificationsService } from 'src/notifications/notifications.service';
 import { CertifcationsService } from './certificates.service';
 import { SaveCertificateDto } from './dto/save-certificate-dto';
 
 @Controller('certificate')
 @ApiTags('certificate')
 export class CertificatesController {
-  constructor(
-    private certificatesService: CertifcationsService,
-    private notificationService: UserNotificationsService,
-  ) {}
+  constructor(private certificatesService: CertifcationsService, private notificationService: NotificationsService) {}
 
   @Get('/:publicId')
   @ApiOperation({ operationId: 'getCertificate' })
