+++ nestjs/src/certificates/certificates.module.ts
@@ -3,12 +3,12 @@ import { Certificate } from '@entities/certificate';
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { ConfigModule } from 'src/config';
+import { NotificationsModule } from 'src/notifications/notifications.module';
 import { CertificatesController } from './certificates.controller';
 import { CertifcationsService } from './certificates.service';
-import { UsersNotificationsModule } from 'src/usersNotifications/usersNotifications.module';
 
 @Module({
-  imports: [TypeOrmModule.forFeature([Certificate, Student]), ConfigModule, UsersNotificationsModule],
+  imports: [TypeOrmModule.forFeature([Certificate, Student]), ConfigModule, NotificationsModule],
   controllers: [CertificatesController],
   providers: [CertifcationsService],
 })
