+++ nestjs/src/notifications/notifications.service.ts
@@ -1,13 +1,15 @@
 import { Notification, NotificationId } from '@entities/notification';
 import { BadRequestException, Injectable, Logger } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
-import { ConfigService } from '../config';
+import { ConfigService } from 'src/config';
 import { Repository } from 'typeorm';
 import { UpdateNotificationDto } from './dto/update-notification.dto';
 import { HttpService } from '@nestjs/axios';
+import { SendNotificationDto } from './dto/send-notification.dto';
 import { EmailTemplate, NotificationChannelSettings, TelegramTemplate } from '@entities/notificationChannelSettings';
 import { compile } from 'handlebars';
 import { NotificationChannelId } from '@entities/notificationChannel';
+import { UserNotificationsService } from 'src/users/users.notifications.service';
 import { emailTemplate } from './email-template';
 import { lastValueFrom } from 'rxjs';
 
@@ -19,8 +21,7 @@ export class NotificationsService {
   constructor(
     @InjectRepository(Notification)
     private notificationsRepository: Repository<Notification>,
-    @InjectRepository(NotificationChannelSettings)
-    private channelSettingsRepository: Repository<NotificationChannelSettings>,
+    private userNotificationsService: UserNotificationsService,
     private configService: ConfigService,
     private httpService: HttpService,
   ) {}
@@ -48,50 +49,33 @@ export class NotificationsService {
   }
 
   /**
-   * Messages to users regarless on user subscription status to specific channel
+   * Automatic user notification based on triggers. sent to subscribed channels based on subscription
    */
-  public async sendMessage(notification: {
-    notificationId: NotificationId;
-    userId: number;
-    data: object;
-    channelId: NotificationChannelId;
-    channelValue: string;
-  }) {
-    const { userId, data, notificationId, channelId, channelValue } = notification;
-    const channelSettings = await this.getChannelSettings(channelId, notificationId);
+  public async sendEventNotification(notification: SendNotificationDto) {
+    const { userId, data, notificationId, expireDate } = notification;
+    const channels = await this.userNotificationsService.getUserNotificationSettings(userId, notificationId);
 
-    const message = this.buildChannelMessage({ ...channelSettings, externalId: channelValue }, data);
+    const channelMap = new Map<NotificationChannelId, NotificationData>();
+    channels.forEach(channel => {
+      const message = this.buildChannelMessage(channel, data);
+      if (message) {
+        const { channelId, template, to } = message;
+        channelMap.set(channelId, { template, to });
+      }
+    });
 
-    if (!message) {
-      this.logger.error({
-        message: `failed to build message fo notification ${notification.notificationId} and user ${notification.userId}`,
-      });
-      return;
-    }
+    if (channelMap.size === 0) return;
 
     await this.publishNotification({
       notificationId,
-      channelId: [channelId],
+      channelId: Array.from(channelMap.keys()),
       userId,
-      data: {
-        [channelId]: {
-          template: message.template,
-          to: message.to,
-        },
-      },
+      expireDate,
+      data: Object.fromEntries(channelMap) as Record<NotificationChannelId, NotificationData>,
     });
   }
 
-  private getChannelSettings(channelId: NotificationChannelId, notificationId: NotificationId) {
-    return this.channelSettingsRepository.findOne({
-      where: {
-        notificationId,
-        channelId,
-      },
-    });
-  }
-
-  public buildChannelMessage(channel: NotificationChannelSettings & { externalId: string }, data: object) {
+  private buildChannelMessage(channel: NotificationChannelSettings & { externalId: string }, data: object) {
     const { channelId, externalId, template } = channel;
     if (!externalId || !template) return;
 
@@ -110,7 +94,7 @@ export class NotificationsService {
     return channelMessage;
   }
 
-  public async publishNotification(notification: NotificationPayload) {
+  private async publishNotification(notification: NotificationPayload) {
     if (this.configService.isDev) return;
 
     const { restApiKey, restApiUrl } = this.configService.awsServices;
@@ -129,10 +113,10 @@ type NotificationPayload = {
   channelId: NotificationChannelId[];
   userId: number;
   expireDate?: number;
-  data: Partial<Record<NotificationChannelId, NotificationData>>;
+  data: Record<NotificationChannelId, NotificationData>;
 };
 
-export type NotificationData = {
+type NotificationData = {
   to: string;
   template: EmailTemplate | TelegramTemplate;
 };
