+++ nestjs/src/profile/profile.module.ts
@@ -3,16 +3,16 @@ import { ProfileService } from './profile.service';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { Course } from '@entities/course';
 import { ProfileController } from './profile.controller';
+import { UsersModule } from '../users/users.module';
 import { CoursesModule } from '../courses/courses.module';
 import { NotificationUserConnection } from '@entities/notificationUserConnection';
 import { User } from '@entities/user';
 import { ProfilePermissions } from '@entities/profilePermissions';
-import { UsersNotificationsModule } from 'src/usersNotifications/usersNotifications.module';
 
 @Module({
   imports: [
     TypeOrmModule.forFeature([Course, NotificationUserConnection, User, ProfilePermissions]),
-    UsersNotificationsModule,
+    UsersModule,
     CoursesModule,
   ],
   controllers: [ProfileController],
