+++ nestjs/src/profile/profile.service.ts
@@ -2,25 +2,20 @@ import { Course } from '@entities/course';
 import { Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { AuthUser } from 'src/auth';
-import { In, Repository, UpdateResult } from 'typeorm';
+import { In, Repository } from 'typeorm';
 import { User } from '@entities/user';
-import { NotificationUserConnection } from '@entities/notificationUserConnection';
 import { ProfilePermissions } from '@entities/profilePermissions';
 import { ProfileInfoDto } from './dto/update-profile.dto';
-import { UserNotificationsService } from 'src/usersNotifications/users.notifications.service';
 
 @Injectable()
 export class ProfileService {
   constructor(
     @InjectRepository(Course)
     private courseRepository: Repository<Course>,
-    @InjectRepository(NotificationUserConnection)
-    private notificationConnectionsRepository: Repository<NotificationUserConnection>,
     @InjectRepository(User)
     private userRepository: Repository<User>,
     @InjectRepository(ProfilePermissions)
     private profilePermissionsRepository: Repository<ProfilePermissions>,
-    private userNotificationsService: UserNotificationsService,
   ) {}
 
   public async getCourses(authUser: AuthUser): Promise<Course[]> {
@@ -60,7 +55,7 @@ export class ProfileService {
       const { skype, phone, email, epamEmail, telegram, notes, linkedIn } = contacts;
       const { countryName, cityName } = location;
 
-      const user = await this.userRepository
+      await this.userRepository
         .createQueryBuilder()
         .update(User)
         .set({
@@ -83,39 +78,6 @@ export class ProfileService {
         .returning('*')
         .where('id = :id', { id: userId })
         .execute();
-
-      await this.updateEmailChannel(userId, user);
-    }
-  }
-
-  private async updateEmailChannel(userId: number, user: UpdateResult) {
-    const newEmail = user.raw[0]?.contactsEmail;
-    const channelId = 'email';
-
-    if (!newEmail) {
-      await this.notificationConnectionsRepository.delete({
-        channelId,
-        userId,
-      });
-    } else {
-      const connection = await this.notificationConnectionsRepository.findOne({
-        where: {
-          channelId,
-          userId,
-        },
-      });
-      const shouldSendEmailConfirmation = !connection || connection.externalId !== newEmail;
-      if (shouldSendEmailConfirmation) {
-        await this.userNotificationsService.sendEmailConfirmation(userId, false);
-      }
-
-      const isConfirmed = connection?.enabled && connection?.externalId === newEmail ? true : false;
-      await this.notificationConnectionsRepository.save({
-        channelId,
-        userId,
-        externalId: newEmail,
-        enabled: isConfirmed,
-      });
     }
   }
 }
