+++ nestjs/src/registry/registry.module.ts
@@ -2,13 +2,13 @@ import { MentorRegistry } from '@entities/mentorRegistry';
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { CoursesModule } from 'src/courses/courses.module';
+import { NotificationsModule } from 'src/notifications/notifications.module';
 import { UsersModule } from 'src/users';
-import { UsersNotificationsModule } from 'src/usersNotifications/usersNotifications.module';
 import { RegistryController } from './registry.controller';
 import { RegistryService } from './registry.service';
 
 @Module({
-  imports: [TypeOrmModule.forFeature([MentorRegistry]), UsersModule, UsersNotificationsModule, CoursesModule],
+  imports: [TypeOrmModule.forFeature([MentorRegistry]), NotificationsModule, UsersModule, CoursesModule],
   controllers: [RegistryController],
   providers: [RegistryService],
 })
