+++ nestjs/src/users/users.module.ts
@@ -2,10 +2,16 @@ import { Module } from '@nestjs/common';
 import { UsersService } from './users.service';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { User } from '@entities/user';
+import { Notification } from '@entities/notification';
+import { UsersNotificationsController } from './users.notifications.controller';
+import { UserNotificationsService } from './users.notifications.service';
+import { NotificationUserConnection } from '@entities/notificationUserConnection';
+import { NotificationUserSettings } from '@entities/notificationUserSettings';
 
 @Module({
-  imports: [TypeOrmModule.forFeature([User])],
-  providers: [UsersService],
-  exports: [UsersService],
+  imports: [TypeOrmModule.forFeature([User, Notification, NotificationUserSettings, NotificationUserConnection])],
+  controllers: [UsersNotificationsController],
+  providers: [UsersService, UserNotificationsService],
+  exports: [UsersService, UserNotificationsService],
 })
 export class UsersModule {}
