--- nestjs/src/usersNotifications/dto/notification-user-connections.dto.ts
@@ -1,20 +1,0 @@
-import { NotificationChannelId } from '@entities/notificationChannel';
-import { ApiProperty } from '@nestjs/swagger';
-import { IsBoolean } from 'class-validator';
-
-export class ConnectionDetails {
-  @ApiProperty()
-  value: string;
-
-  @ApiProperty()
-  @IsBoolean()
-  enabled: boolean;
-
-  @ApiProperty()
-  lastLinkSentAt: string;
-}
-
-export class NotificationUserConnectionsDto {
-  @ApiProperty({ type: Map })
-  public connections: Record<NotificationChannelId, ConnectionDetails>;
-}
