+++ server/src/models/loginState.ts
@@ -10,16 +10,8 @@ export class LoginState {
   @Index()
   createdDate: number;
 
-  @Index()
-  @Column({ nullable: true })
-  userId: number;
-
   @Column({ type: 'simple-json' })
   data: LoginData;
-
-  @Index()
-  @Column({ type: 'timestamp', nullable: true })
-  expires?: string;
 }
 
 export type LoginData = Partial<{
