+++ server/src/models/notificationUserConnection.ts
@@ -1,18 +1,8 @@
-import {
-  Entity,
-  Column,
-  CreateDateColumn,
-  PrimaryColumn,
-  ManyToOne,
-  JoinColumn,
-  UpdateDateColumn,
-  Unique,
-} from 'typeorm';
+import { Entity, Column, CreateDateColumn, PrimaryColumn, ManyToOne, JoinColumn, UpdateDateColumn } from 'typeorm';
 import { NotificationChannel, User } from '.';
 import { NotificationChannelId } from './notificationChannel';
 
 @Entity()
-@Unique(['userId', 'channelId', 'externalId'])
 export class NotificationUserConnection {
   @ManyToOne(() => User, user => user.notificationConnections, { onDelete: 'CASCADE', onUpdate: 'CASCADE' })
   @JoinColumn()
@@ -33,7 +23,7 @@ export class NotificationUserConnection {
   @PrimaryColumn()
   channelId: NotificationChannelId;
 
-  @Column()
+  @PrimaryColumn()
   externalId: string;
 
   @Column({ default: true })
