+++ server/src/services/notification.service.ts
@@ -4,7 +4,7 @@ import { config } from '../config';
 export async function sendNotification(notification: NotificationV2) {
   const { password, username } = config.users.cloud;
 
-  await axios.post(`${config.host}/api/v2/users/notifications/send`, notification, {
+  await axios.post(`${config.host}/api/v2/notifications/send`, notification, {
     headers: {
       Authorization: `Basic ${Buffer.from(username + ':' + password).toString('base64')}`,
     },
